# MTLS_UI/src/openapi/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/options": {
            "post": {
                "operationId": "OptionController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOptionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tùy chọn đã được tạo thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                },
                "summary": "Tạo tùy chọn mới cho câu hỏi",
                "tags": ["options"]
            },
            "get": {
                "operationId": "OptionController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Trả về danh sách tùy chọn"
                    }
                },
                "summary": "Lấy danh sách tất cả các tùy chọn",
                "tags": ["options"]
            }
        },
        "/options/{id}": {
            "get": {
                "operationId": "OptionController_findOne",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của tùy chọn",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về thông tin tùy chọn"
                    },
                    "404": {
                        "description": "Không tìm thấy tùy chọn"
                    }
                },
                "summary": "Lấy thông tin tùy chọn theo ID",
                "tags": ["options"]
            },
            "put": {
                "operationId": "OptionController_update",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của tùy chọn",
                    "schema": {
                        "type": "string"
                    }
                }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateOptionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tùy chọn đã được cập nhật"
                    },
                    "404": {
                        "description": "Không tìm thấy tùy chọn"
                    }
                },
                "summary": "Cập nhật tùy chọn",
                "tags": ["options"]
            },
            "delete": {
                "operationId": "OptionController_remove",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của tùy chọn",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Tùy chọn đã được xóa"
                    },
                    "404": {
                        "description": "Không tìm thấy tùy chọn"
                    }
                },
                "summary": "Xóa tùy chọn",
                "tags": ["options"]
            }
        },
        "/options/question/{id}": {
            "get": {
                "operationId": "OptionController_findListOptionByQuestionId",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của câu hỏi",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về danh sách tùy chọn"
                    },
                    "404": {
                        "description": "Không tìm thấy tùy chọn"
                    }
                },
                "summary": "Lấy danh sách tùy chọn theo ID của câu hỏi",
                "tags": ["options"]
            }
        },
        "/options/batch": {
            "post": {
                "operationId": "OptionController_createMultiple",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateMultipleOptionsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Các tùy chọn đã được tạo thành công"
                    }
                },
                "summary": "Tạo nhiều tùy chọn cùng lúc",
                "tags": ["options"]
            }
        },
...
        "/notes": {
            "post": {
                "operationId": "NoteController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateNoteDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ghi chú đã được tạo thành công"
                    }
                },
                "security": [{
                    "bearer": []
                }],
                "summary": "Tạo ghi chú mới cho module",
                "tags": ["notes"]
            },
            "get": {
                "operationId": "NoteController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Trả về danh sách ghi chú"
                    }
                },
                "summary": "Lấy danh sách tất cả ghi chú",
                "tags": ["notes"]
            }
        },
        "/notes/{id}": {
            "get": {
                "operationId": "NoteController_findOne",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của ghi chú",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về thông tin ghi chú"
                    },
                    "404": {
                        "description": "Không tìm thấy ghi chú"
                    }
                },
                "summary": "Lấy thông tin ghi chú theo ID",
                "tags": ["notes"]
            },
            "patch": {
                "operationId": "NoteController_update",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của ghi chú",
                    "schema": {
                        "type": "string"
                    }
                }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateNoteDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ghi chú đã được cập nhật"
                    },
                    "404": {
                        "description": "Không tìm thấy ghi chú hoặc không có quyền"
                    }
                },
                "security": [{
                    "bearer": []
                }],
                "summary": "Cập nhật ghi chú",
                "tags": ["notes"]
            },
            "delete": {
                "operationId": "NoteController_remove",
                "parameters": [{
                    "name": "id",
                    "required": true,
                    "in": "path",
                    "description": "ID của ghi chú",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Ghi chú đã được xóa"
                    },
                    "404": {
                        "description": "Không tìm thấy ghi chú hoặc không có quyền"
                    }
                },
                "security": [{
                    "bearer": []
                }],
                "summary": "Xóa ghi chú",
                "tags": ["notes"]
            }
        },
        "/notes/student/me": {
            "get": {
                "operationId": "NoteController_findMyNotes",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Trả về danh sách ghi chú"
                    }
                },
                "security": [{
                    "bearer": []
                }],
                "summary": "Lấy danh sách ghi chú của học viên đang đăng nhập",
                "tags": ["notes"]
            }
        },
        "/notes/student/{studentId}": {
            "get": {
                "operationId": "NoteController_findByStudentId",
                "parameters": [{
                    "name": "studentId",
                    "required": true,
                    "in": "path",
                    "description": "ID của học viên",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về danh sách ghi chú"
                    }
                },
                "summary": "Lấy danh sách ghi chú của học viên theo ID",
                "tags": ["notes"]
            }
        },
        "/notes/module/{moduleId}": {
            "get": {
                "operationId": "NoteController_findByModuleId",
                "parameters": [{
                    "name": "moduleId",
                    "required": true,
                    "in": "path",
                    "description": "ID của module",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Trả về danh sách ghi chú"
                    }
                },
                "summary": "Lấy danh sách ghi chú của module theo ID",
                "tags": ["notes"]
            }
        },
        "/notes/student/{studentId}/module/{moduleId}": {
            "get": {
                "operationId": "NoteController_findByStudentAndModule",
                "parameters": [{
                        "name": "studentId",
                        "required": true,
                        "in": "path",
                        "description": "ID của học viên",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "moduleId",
                        "required": true,
                        "in": "path",
                        "description": "ID của module",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trả về danh sách ghi chú"
                    }
                },
                "summary": "Lấy danh sách ghi chú của học viên cho một module cụ thể",
                "tags": ["notes"]
            }
        },
...
    },
    "info": {
        "title": "API Documentation",
        "description": "The API's MTLS description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateOptionDto": {
                "type": "object",
                "properties": {
                    "question": {
                        "type": "string",
                        "example": "60d21b4667d0d8992e610c85"
                    },
                    "content": {
                        "type": "string",
                        "example": "Option content here"
                    },
                    "isCorrect": {
                        "type": "boolean",
                        "example": true
                    },
                    "isActive": {
                        "type": "boolean",
                        "description": "Trạng thái hoạt động của tùy chọn",
                        "example": true,
                        "default": true
                    }
                },
                "required": ["question", "content", "isCorrect"]
            },
            "UpdateOptionDto": {
                "type": "object",
                "properties": {
                    "question": {
                        "type": "string",
                        "example": "60d21b4667d0d8992e610c85"
                    },
                    "content": {
                        "type": "string",
                        "example": "Updated option content here"
                    },
                    "isCorrect": {
                        "type": "boolean",
                        "example": false
                    }
                }
            },
            "CreateMultipleOptionsDto": {
                "type": "object",
                "properties": {
                    "options": {
                        "description": "Danh sách các tùy chọn cần tạo",
                        "example": [{
                                "question": "60d21b4667d0d8992e610c85",
                                "content": "A minor",
                                "isCorrect": true
                            },
                            {
                                "question": "60d21b4667d0d8992e610c85",
                                "content": "B minor",
                                "isCorrect": false
                            }
                        ],
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CreateOptionDto"
                        }
                    }
                },
                "required": ["options"]
            },
...
            "CreateNoteDto": {
                "type": "object",
                "properties": {
                    "module": {
                        "type": "string",
                        "description": "ID của module mà note thuộc về",
                        "example": "6507d6f6c31c7948c831a123"
                    },
                    "student": {
                        "type": "string",
                        "description": "ID của học viên (sẽ được lấy từ token)",
                        "example": "6507d6f6c31c7948c831b789"
                    },
                    "content": {
                        "type": "string",
                        "description": "Nội dung ghi chú",
                        "example": "Cần nhớ về các dấu chấm nhịp trong âm nhạc"
                    }
                },
                "required": ["module", "content"]
            },
            "UpdateNoteDto": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "string",
                        "description": "Nội dung ghi chú",
                        "example": "Cập nhật: Cần nhớ về các dấu chấm nhịp và dấu lặng trong âm nhạc"
                    }
                }
            },
...
        }
    }
}